requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
Faker==23.1.0
hugchat==0.3.11
idna==3.6
isoweek==1.3.3
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
whitenoise==6.6.0
xlwt==1.3.0
